refactor: name 클래스 분리로 인한 테스트 코드 추가 및 수정
